+++ app/src/main/java/com/foke/together/BaseApplication.kt
@@ -1,0 +1,8 @@
+package com.foke.together
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+
+@HiltAndroidApp
+class BaseApplication:Application()
