+++ app/src/main/java/com/foke/together/data/repository/CustomCameraRepoImpl.kt
@@ -1,0 +1,100 @@
+package com.foke.together.data.repository
+
+import android.content.ContentValues
+import android.content.Context
+import android.os.Build
+import android.provider.MediaStore
+import android.widget.Toast
+import androidx.camera.core.*
+import androidx.camera.lifecycle.ProcessCameraProvider
+import androidx.camera.view.PreviewView
+import androidx.core.content.ContextCompat
+import androidx.lifecycle.LifecycleOwner
+import com.foke.together.domain.repository.CustomCameraRepo
+import java.text.SimpleDateFormat
+import java.util.*
+import javax.inject.Inject
+
+class CustomCameraRepoImpl @Inject constructor(
+    private val cameraProvider: ProcessCameraProvider,
+    private val selector: CameraSelector,
+    private val preview: Preview,
+    private val imageAnalysis: ImageAnalysis,
+    private val imageCapture: ImageCapture
+):CustomCameraRepo {
+
+
+    override suspend fun captureAndSaveImage(
+        context: Context) {
+
+        //for file name
+        val name = SimpleDateFormat(
+            "yyyy-MM-dd-HH-mm-ss-SSS",
+            Locale.ENGLISH
+        ).format(System.currentTimeMillis())
+
+
+        // for storing
+        val contentValues = ContentValues().apply {
+            put(MediaStore.MediaColumns.DISPLAY_NAME,name)
+            put(MediaStore.MediaColumns.MIME_TYPE,"image/jpeg")
+            put(MediaStore.Images.Media.RELATIVE_PATH,"Pictures/My-Camera-App-Images")
+        }
+
+        // for capture output
+        val outputOptions = ImageCapture.OutputFileOptions
+            .Builder(
+                context.contentResolver,
+                MediaStore.Images.Media.EXTERNAL_CONTENT_URI,
+                contentValues
+            )
+            .build()
+
+        imageCapture.takePicture(
+            outputOptions,
+            ContextCompat.getMainExecutor(context),
+            object : ImageCapture.OnImageSavedCallback{
+                override fun onImageSaved(outputFileResults: ImageCapture.OutputFileResults) {
+                   Toast.makeText(
+                       context,
+                       "Saved image ${outputFileResults.savedUri!!}",
+                       Toast.LENGTH_LONG
+                   ).show()
+                }
+
+                override fun onError(exception: ImageCaptureException) {
+                    Toast.makeText(
+                        context,
+                        "some error occurred ${exception.message}",
+                        Toast.LENGTH_LONG
+                    ).show()
+                }
+            }
+        )
+
+    }
+
+
+
+
+    override suspend fun showCameraPreview(
+        previewView: PreviewView,
+        lifecycleOwner: LifecycleOwner
+    ) {
+
+        preview.setSurfaceProvider(previewView.surfaceProvider)
+        try {
+            cameraProvider.unbindAll()
+            cameraProvider.bindToLifecycle(
+                lifecycleOwner,
+                selector,
+                preview,
+                imageAnalysis,
+                imageCapture
+            )
+        }catch (e:Exception){
+            e.printStackTrace()
+        }
+    }
+
+}
