+++ app/src/main/java/com/foke/together/di/AppModule.kt
@@ -1,0 +1,78 @@
+package com.foke.together.di
+
+import android.app.Application
+import androidx.camera.core.*
+import androidx.camera.core.ImageAnalysis.STRATEGY_KEEP_ONLY_LATEST
+import androidx.camera.core.ImageCapture.FLASH_MODE_ON
+import androidx.camera.lifecycle.ProcessCameraProvider
+import com.foke.together.data.repository.CustomCameraRepoImpl
+import com.foke.together.domain.repository.CustomCameraRepo
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+
+@Module
+@InstallIn(SingletonComponent::class)
+object AppModule {
+
+    @Provides
+    @Singleton
+    fun provideCameraSelector(): CameraSelector {
+        return CameraSelector.Builder()
+            .requireLensFacing(CameraSelector.LENS_FACING_BACK)
+            .build()
+    }
+
+    @Provides
+    @Singleton
+    fun provideCameraProvider(application: Application)
+            : ProcessCameraProvider {
+        return ProcessCameraProvider.getInstance(application).get()
+
+    }
+
+    @Provides
+    @Singleton
+    fun provideCameraPreview(): Preview {
+        return Preview.Builder().build()
+    }
+
+    @Provides
+    @Singleton
+    fun provideImageCapture(): ImageCapture {
+        return ImageCapture.Builder()
+            .setFlashMode(FLASH_MODE_ON)
+            .setTargetAspectRatio(AspectRatio.RATIO_16_9)
+            .build()
+    }
+
+    @Provides
+    @Singleton
+    fun provideImageAnalysis(): ImageAnalysis {
+        return ImageAnalysis.Builder()
+            .setBackpressureStrategy(STRATEGY_KEEP_ONLY_LATEST)
+            .build()
+    }
+
+    @Provides
+    @Singleton
+    fun provideCustomCameraRepo(
+        cameraProvider: ProcessCameraProvider,
+        selector: CameraSelector,
+        imageCapture: ImageCapture,
+        imageAnalysis: ImageAnalysis,
+        preview: Preview
+    ): CustomCameraRepo {
+        return CustomCameraRepoImpl(
+            cameraProvider,
+            selector,
+            preview,
+            imageAnalysis,
+            imageCapture
+        )
+    }
+
+}
