+++ app/src/main/java/com/foke/together/domain/repository/CustomCameraRepo.kt
@@ -1,0 +1,14 @@
+package com.foke.together.domain.repository
+
+import android.content.Context
+import androidx.camera.view.PreviewView
+import androidx.lifecycle.LifecycleOwner
+
+interface CustomCameraRepo {
+
+    suspend fun captureAndSaveImage(context:Context)
+    suspend fun showCameraPreview(
+        previewView: PreviewView,
+        lifecycleOwner: LifecycleOwner
+    )
+}
