+++ app/src/main/java/com/foke/together/presentation/MainScreen.kt
@@ -1,0 +1,34 @@
+package com.foke.together.presentation
+
+import android.Manifest
+import android.os.Build
+import android.widget.Toast
+import androidx.camera.view.PreviewView
+import androidx.compose.foundation.layout.*
+import androidx.compose.material.Icon
+import androidx.compose.material.IconButton
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.SideEffect
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalLifecycleOwner
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.viewinterop.AndroidView
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.navigation.compose.rememberNavController
+import com.google.accompanist.permissions.ExperimentalPermissionsApi
+import com.google.accompanist.permissions.rememberMultiplePermissionsState
+import com.foke.together.R
+import com.foke.together.presentation.navigation.NavGraph
+
+
+@OptIn(ExperimentalPermissionsApi::class)
+@Composable
+fun MainScreen() {
+    val navController = rememberNavController()
+    NavGraph(navController)
+}
