+++ app/src/main/java/com/foke/together/presentation/MainViewModel.kt
@@ -1,0 +1,39 @@
+package com.foke.together.presentation
+
+import android.content.Context
+import androidx.camera.core.Preview
+import androidx.camera.view.PreviewView
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.foke.together.domain.repository.CustomCameraRepo
+import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.launch
+import javax.inject.Inject
+
+
+@HiltViewModel
+class MainViewModel @Inject constructor(
+    private val repo: CustomCameraRepo
+):ViewModel() {
+
+    fun showCameraPreview(
+        previewView: PreviewView,
+        lifecycleOwner: LifecycleOwner
+    ){
+        viewModelScope.launch {
+            repo.showCameraPreview(
+                previewView,
+                lifecycleOwner
+            )
+        }
+    }
+
+    fun captureAndSave(context: Context){
+        viewModelScope.launch {
+            repo.captureAndSaveImage(context)
+        }
+    }
+
+
+}
