+++ app/src/main/java/com/foke/together/presentation/screen/FrameScreen.kt
@@ -1,0 +1,11 @@
+package com.foke.together.presentation.screen
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun FrameScreen(
+    popBackStack: () -> Unit,
+    navigateToHome:() -> Unit
+) {
+
+}
