+++ app/src/main/java/com/foke/together/presentation/screen/SettingScreen.kt
@@ -1,0 +1,11 @@
+package com.foke.together.presentation.screen
+
+import androidx.compose.runtime.Composable
+import androidx.hilt.navigation.compose.hiltViewModel
+
+@Composable
+fun SettingScreen(
+    popBackStack: () -> Unit
+) {
+
+}
