--- app/src/main/java/com/kamatiakash/customcameraappwithjetpackcompose/BaseApplication.kt
@@ -1,8 +1,0 @@
-package com.kamatiakash.customcameraappwithjetpackcompose
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-
-@HiltAndroidApp
-class BaseApplication:Application()
