--- app/src/main/java/com/kamatiakash/customcameraappwithjetpackcompose/domain/repository/CustomCameraRepo.kt
@@ -1,14 +1,0 @@
-package com.kamatiakash.customcameraappwithjetpackcompose.domain.repository
-
-import android.content.Context
-import androidx.camera.view.PreviewView
-import androidx.lifecycle.LifecycleOwner
-
-interface CustomCameraRepo {
-
-    suspend fun captureAndSaveImage(context:Context)
-    suspend fun showCameraPreview(
-        previewView: PreviewView,
-        lifecycleOwner: LifecycleOwner
-    )
-}
