--- app/src/main/java/com/kamatiakash/customcameraappwithjetpackcompose/presentation/MainViewModel.kt
@@ -1,39 +1,0 @@
-package com.kamatiakash.customcameraappwithjetpackcompose.presentation
-
-import android.content.Context
-import androidx.camera.core.Preview
-import androidx.camera.view.PreviewView
-import androidx.lifecycle.LifecycleOwner
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.kamatiakash.customcameraappwithjetpackcompose.domain.repository.CustomCameraRepo
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.launch
-import javax.inject.Inject
-
-
-@HiltViewModel
-class MainViewModel @Inject constructor(
-    private val repo: CustomCameraRepo
-):ViewModel() {
-
-    fun showCameraPreview(
-        previewView: PreviewView,
-        lifecycleOwner: LifecycleOwner
-    ){
-        viewModelScope.launch {
-            repo.showCameraPreview(
-                previewView,
-                lifecycleOwner
-            )
-        }
-    }
-
-    fun captureAndSave(context: Context){
-        viewModelScope.launch {
-            repo.captureAndSaveImage(context)
-        }
-    }
-
-
-}
