--- app/src/main/java/com/kamatiakash/customcameraappwithjetpackcompose/presentation/screen/FrameScreen.kt
@@ -1,11 +1,0 @@
-package com.kamatiakash.customcameraappwithjetpackcompose.presentation.screen
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun FrameScreen(
-    popBackStack: () -> Unit,
-    navigateToHome:() -> Unit
-) {
-
-}
