--- app/src/main/java/com/kamatiakash/customcameraappwithjetpackcompose/presentation/screen/SettingScreen.kt
@@ -1,11 +1,0 @@
-package com.kamatiakash.customcameraappwithjetpackcompose.presentation.screen
-
-import androidx.compose.runtime.Composable
-import androidx.hilt.navigation.compose.hiltViewModel
-
-@Composable
-fun SettingScreen(
-    popBackStack: () -> Unit
-) {
-
-}
