+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Custom camera app with jetpack compose</string>
+    <string name="app_name">FourCuts Together</string>
 </resources>
